+++ HISTORY.rst
@@ -10,6 +10,7 @@ Pending
 * Now MySQL 5.7 compatible
 * The final message from ``SmartChunkedIterator`` is now rounded to the nearest
   second.
+* Lock supports lock / unlock methods
 
 1.0.5 (2016-02-10)
 ------------------